+++ ft_memset.c
@@ -10,10 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	size;
-	unsigned char		*buff;
+	size_t	size;
+	unsigned char	*buff;
 
 	if (!s)
 		return (s);
